refactor : CommandController 이름 수정 & model 경로 수정
