+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/Voucher.java
@@ -1,4 +1,4 @@
-package com.example.commandlineapplication.domain.voucher.model;
+package com.example.commandlineapplication.domain.voucher;
 
 import java.util.UUID;
