+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/dto/mapper/VoucherMapper.java
@@ -1,9 +1,9 @@
 package com.example.commandlineapplication.domain.voucher.dto.mapper;
 
+import com.example.commandlineapplication.domain.voucher.Voucher;
+import com.example.commandlineapplication.domain.voucher.VoucherType;
 import com.example.commandlineapplication.domain.voucher.dto.request.VoucherCreateRequest;
 import com.example.commandlineapplication.domain.voucher.dto.response.VoucherResponse;
-import com.example.commandlineapplication.domain.voucher.model.Voucher;
-import com.example.commandlineapplication.domain.voucher.model.VoucherType;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
