+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/dto/request/VoucherCreateRequest.java
@@ -1,6 +1,6 @@
 package com.example.commandlineapplication.domain.voucher.dto.request;
 
-import com.example.commandlineapplication.domain.voucher.model.VoucherType;
+import com.example.commandlineapplication.domain.voucher.VoucherType;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
