+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/repository/VoucherMemoryRepository.java
@@ -1,6 +1,6 @@
 package com.example.commandlineapplication.domain.voucher.repository;
 
-import com.example.commandlineapplication.domain.voucher.model.Voucher;
+import com.example.commandlineapplication.domain.voucher.Voucher;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
