+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/repository/VoucherRepository.java
@@ -1,6 +1,6 @@
 package com.example.commandlineapplication.domain.voucher.repository;
 
-import com.example.commandlineapplication.domain.voucher.model.Voucher;
+import com.example.commandlineapplication.domain.voucher.Voucher;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
