+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/service/VoucherService.java
@@ -1,10 +1,10 @@
 package com.example.commandlineapplication.domain.voucher.service;
 
+import com.example.commandlineapplication.domain.voucher.Voucher;
+import com.example.commandlineapplication.domain.voucher.VoucherType;
 import com.example.commandlineapplication.domain.voucher.dto.mapper.VoucherMapper;
 import com.example.commandlineapplication.domain.voucher.dto.request.VoucherCreateRequest;
 import com.example.commandlineapplication.domain.voucher.dto.response.VoucherResponse;
-import com.example.commandlineapplication.domain.voucher.model.Voucher;
-import com.example.commandlineapplication.domain.voucher.model.VoucherType;
 import com.example.commandlineapplication.domain.voucher.repository.VoucherRepository;
 import java.util.List;
 import java.util.UUID;
